Refactor plot_compilation_size
